requirements.txt
aiohttp
fake-useragent
pyyaml
prometheus-client
selenium
pandas
web3
plotly
python-dotenv